% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sitecode_column.R
\name{add_sitecode_column}
\alias{add_sitecode_column}
\title{Add AQS_SITECODE column to dataframe}
\usage{
add_sitecode_column(df)
}
\arguments{
\item{df}{A data frame with columns state_code, county_code, and site_number.}
}
\value{
data.frame
}
\description{
Adds a column names aqs_sitecode to a dataframe by concatenating the
state_code, county_code, and site_number. The dataframe must contain these
named columns, each of type character.
}
\examples{
df <- data.frame(state_code = '32', county_code = '003',
                 site_number = '0002')
add_sitecode_column(df)
}
