% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_annual_poc.R
\name{adjust_annual_poc}
\alias{adjust_annual_poc}
\title{Adjust POC value for summaries of annual data}
\usage{
adjust_annual_poc(annual_data)
}
\arguments{
\item{annual_data}{data.frame Annual data data.frame.}
}
\value{
data.frame Annual data with adjusted POC values
}
\description{
Where multiple POCs exist for a given site, parameter, method, year,
set all POC ids to the minimum for that site to calculate summaries
}
\examples{
adjust_annual_poc(annual_data)
}
