% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_trend_data.R
\name{annual_trend_data}
\alias{annual_trend_data}
\title{Generate annual trend data}
\usage{
annual_trend_data(tidy_annual_data, .agg_col = arithmetic_mean, ...)
}
\arguments{
\item{tidy_annual_data}{data.frame Annual data that has ideally been
passed through the
\code{tidy_annual_data()} function.}

\item{.agg_col}{key Column to calculated statistics for (unquoted).}

\item{...}{key Columns to group annual data by (unquoted).}
}
\value{
data.frame Annual trend data
}
\description{
Returns a dataframe with the mean annual value the specified .agg_col
by year and any other grouping columns passed to the ... argument.
}
\details{
Takes in annual data data.frame,
and  (optionally) passed through the \code{tidy_annual_data()} function to filter
data to only aqs_sitecode-parameter-method-poc-sample_duration combinations
with data from >= 75\% of all years.
}
\examples{
tidy_annual_data <- tidy_annual_data(annual_data)
annual_trend_data(tidy_annual_data,
                  .agg_col = arithmetic_mean,
                  aqs_sitecode, parameter, method,
                  sample_duration, pollutant_standard)
}
