% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ymd-class.R, R/pad2-class.R,
%   R/pad3-class.R, R/pad4-class.R, R/pad9-class.R
\name{as}
\alias{as}
\title{As("character", "date_ymd")}
\description{
As("character", "date_ymd")

As("character", "pad2")

As("character", "pad3")

As("character", "pad4")

As("character", "pad9")
}
\seealso{
Other date_ymd: 
\code{\link{date_ymd-class}}

Other pad2: 
\code{\link{pad2-class}}

Other pad3: 
\code{\link{pad3-class}}

Other pad4: 
\code{\link{pad4-class}}

Other pad9: 
\code{\link{pad9-class}}
}
\concept{date_ymd}
\concept{pad2}
\concept{pad3}
\concept{pad4}
\concept{pad9}
