% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_col_formats.R
\name{assign_col_formats}
\alias{assign_col_formats}
\title{Parse columns with possible leading zeros}
\usage{
assign_col_formats(path, files_to_import, subfolder = "")
}
\arguments{
\item{path}{string Filepath}

\item{files_to_import}{list List of file names}
}
\value{
list Lists of columns to be formatted as each custom class
}
\description{
Retrives the column names for each type of service file and parses
existing columns into formatting groups (see set_column_classes) and columns
that are unused in the codes (drop columns).
Returns a list of lists, where each nested list is a set of filenames
for each format type in the following order:
list(pad2_cols, pad3_cols, pad4_cols, pad9_cols, date_cols, drop_cols)
}
\details{
The following columns, if they exist in a file, are assigned to each group:
\itemize{
\item pad2 = state_code
\item pad3 = method_code, method code, county_code in any case
\item pad4 = site_number, agency code, agency_code, monitoring_agency_code,
pqao_code in any case
\item pad9 = aqs_site_id
\item date_cols = assessment_date, select_date, date_local
\item drop_cols = columns from full list of unused columns that exist in file
}
}
\keyword{internal}
