% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bias.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate bias value}
\usage{
calculate_bias(data, .pct_diff, ...)
}
\arguments{
\item{data}{data.frame Data with a column for percent difference}

\item{.pct_diff}{key The unquoted column name of the column with percent
difference values}

\item{...}{key Unquoted column names of columns to group by.}
}
\value{
data.frame Bias and count values per grouping
}
\description{
Calculates the bias within a grouping of data. Percent difference value
must already be calculated as a column in the data.
}
\details{
Groupings may be specified in the ... argument.
}
\examples{
calculate_bias(one_point_qc,
               .pct_diff = percent_difference,
               method,
               pqao)
}
