% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cv.R
\name{calculate_cv}
\alias{calculate_cv}
\title{Calculate CV value}
\usage{
calculate_cv(data, .pct_diff, ...)
}
\arguments{
\item{data}{data.frame Data with a column for percent difference}

\item{.pct_diff}{key The unquoted column name of the column with percent
difference values}

\item{...}{key Unquoted column names of columns to group by.}
}
\value{
data.frame CV and count values per grouping
}
\description{
Calculates the CVs within a grouping of data. Percent difference value
must already be calculated/available as a column in the data.
}
\details{
Groupings may be specified in the ... argument.
}
\examples{
calculate_cv(one_point_qc,
             .pct_diff = percent_difference,
             method,
             pqao)
}
