% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_codes.R
\name{clean_codes}
\alias{clean_codes}
\title{Fixes leading zeros}
\usage{
clean_codes(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
data.frame
}
\description{
Fixes leading zeros issues tribal, pqao and agency
codes. Fills NA pqao/pqao_code with monitoring_agency/code.
Any of the following present columns will be padded with zeros:
- tribal_code (3)
- monitoring_agency_code (3/4) depending of tribal status
- pqao_code (3/4) depending of tribal status
}
\keyword{internal}
