% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness_data.R
\name{completeness_data}
\alias{completeness_data}
\title{Generate completeness data}
\usage{
completeness_data(annual_data_adj_poc, required_param_codes = NULL)
}
\arguments{
\item{annual_data_adj_poc}{data.frame Annual data with POC adjusted by
through \code{adjust_annual_poc()}.}

\item{required_param_codes}{character vector (optional) List of required
parameters to filter results
by.}
}
\value{
data.frame Site and site-parameter completeness data
}
\description{
Returns dataframe with site completeness metrics, and site-parameter
completeness metrics.
}
\details{
Requires annual_data data.frame. Optionally accepts
a character vector of required parameters to filter by.
}
\examples{
annual_data_adj_poc <- adjust_annual_poc(annual_data)
network_completeness_data <- completeness_data(annual_data_adj_poc)
network_completeness_data <- completeness_data(annual_data_adj_poc, c('42101'))
}
