% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_year_df.R
\name{create_year_df}
\alias{create_year_df}
\title{Create year_DF dataframe}
\usage{
create_year_df(year_master)
}
\arguments{
\item{year_master}{data.frame The master year list, with column 'year'
all years in the data}
}
\value{
data.frame year_DF with columns start/end_year, start/end_date

data.frame
}
\description{
Create year_DF dataframe
}
