% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_files_by_date.R
\name{filter_files_by_date}
\alias{filter_files_by_date}
\title{Filter files by date range}
\usage{
filter_files_by_date(date_range, files_to_import)
}
\arguments{
\item{date_range}{character vector A start date/year (string) or date/year
range (vector) in the format 'yyyymmdd' or
'yyyy'. If range given, vector should have
two elements, c(start_date/year,
end_date/year)
If only start date/year given, all files
with data >= start date will be included.}

\item{files_to_import}{character vector A list of filenames in the format
'ppppp_yyyymmdd_yyyymmdd_service.*'}
}
\value{
character Subest of files_to_import that meets specified criteria.
}
\description{
From a list of filenames in the format
ppppp_yyyymmdd_yyyymmdd_service
- first date is the begin-date of the data
- second date is the end-date of the data
- ppppp is a parameter code
returns the files that contain data within a specified year or
date range.
}
\keyword{internal}
