% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_dv_sitelist.R
\name{master_dv_sitelist}
\alias{master_dv_sitelist}
\title{Create maser DV Sitelist}
\usage{
master_dv_sitelist(dv_df, site_master, duration, dv_yr_prev)
}
\arguments{
\item{dv_df}{data.frame The design values of interest imported with the
import_design_values() function}

\item{site_master}{data.frame Site master dataframe created with the
master_combination_list() function}

\item{duration}{string The duration of design values input as dv_df. Options
include '8hr', '1hr', '24hr', 'annual'}

\item{dv_yr_prev}{integer The number of years previous to the most recent
available in dv_df to use for design values (i.e.
if 2019 is the most recent, dv_yr_prev = 2 would
select the 2017 design values; dv_yr_prev = 0 would
select the 2019 design values)}
}
\value{
data.frame
}
\description{
Create maser DV Sitelist
}
