% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_method_list.R
\name{master_method_list}
\alias{master_method_list}
\title{Create master pqao method list}
\usage{
master_method_list(df_vec)
}
\arguments{
\item{df_vec}{vector Collection of dataframes to include in creating of master
pqao-method list. Dfs should have columns "pqao",
"method", "method_code", "method_type",
"equivalent_method" (see tidy_method() function).}
}
\value{
data.frame The master pqao-method list
}
\description{
Requested df's should have the columns "pqao", "method", "method_code",
"method_type", "equivalent_method". The latter two columns are joined
from the methods_list (metadata) dataframe via the tidy_method() function.
}
