% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mdl_summary.R
\name{method_mdl_summary}
\alias{method_mdl_summary}
\title{Method MDL Summary Data}
\usage{
method_mdl_summary(sample_data)
}
\arguments{
\item{sample_data}{data.frame Sample data as read in by \code{import_raw_data()}}
}
\value{
data.frame Summary statistics about detection limits for hourly data
}
\description{
Summary of detection limit stats per method, including:
\itemize{
\item Mean, median, min, max detection limit per method
\item \% of data above detection limit per parameter/method
}
}
\examples{
method_mdl_summary(sample_data)
}
