% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution_rose_data.R
\name{pollution_rose_data}
\alias{pollution_rose_data}
\title{Pollution Rose Data}
\usage{
pollution_rose_data(
  sample_data,
  wind_data,
  param_code1,
  duration = "1 HOUR",
  param_code2 = NULL
)
}
\arguments{
\item{sample_data}{data.frame Sample data data.frame}

\item{wind_data}{data.frame Wind data as generated by \code{wind_data()}}

\item{param_code1}{string Parameter code to include in dataframe}

\item{duration}{string Sample duration to filter to, as listed in the
\code{sample_duration} column}

\item{param_code2}{string (optional) Second parameter code to include in
dataframe}
}
\value{
data.frame Tidied dataframe with all data needed to create a pollution
rose.
}
\description{
Requires sample data (as read-in by \code{import_raw_data()}) and wind data
as generated by \code{wind_data()}.
}
\details{
Returns a dataframe with a column of measurements per specified parameter
(up to two), a column for wind data, a column for site code, and a column
for datetime.

Columns are added for year and season, where seasons are:
\itemize{
\item Winter: 12, 1, 2
\item Spring: 3, 4, 5
\item Summer: 6, 7, 8
\item Fall: 9, 10, 11
}

Generates dataframe to pass to pollutionRoseServer() function in
sonomaDashboardUI package...
}
