% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_method.R
\name{prep_method}
\alias{prep_method}
\title{Join methods information to requested df(s)}
\usage{
prep_method(methods_list, df)
}
\arguments{
\item{methods_list}{data.frame The methods_list imported with import_metadata}

\item{df}{data.frame A data frame to add methods to. Should have columns
"method code", 'parameter'.}
}
\value{
data.frame
}
\description{
Joins method metadata information to df(s) requested in
name_df_list and re-assigns updated df to variable of
same name. Dfs should have columns "method_code", 'parameter',
'parameter_code'.
}
\details{
Drops rows with NA in method_code
}
\examples{
prep_method(methods_list, one_point_qc)
}
