% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pqao.R
\name{prep_pqao}
\alias{prep_pqao}
\title{Prepare pqao values in qc and raw data}
\usage{
prep_pqao(pqao_list, monitors_list, qc_data, df, df_type)
}
\arguments{
\item{pqao_list}{data.frame The pqao_list imported with import_metadata}

\item{monitors_list}{data.frame The monitors_list metadata}

\item{qc_data}{data.frame QC dataframe#'}

\item{df}{data.frame A data frame to join pqao data to.}

\item{df_type}{string Either 'raw' for annual, sample, or daily data, or
'qc' for any qc data.}
}
\value{
data.frame
}
\description{
Creates full list of pqaos by combining qc data and monitors data.
Adds agency type column from pqao_list
Updates pqaos in qc data.
Joins pqaos to raw data.
}
\examples{
prep_pqao(pqao_list, monitors_list, one_point_qc, daily_data, 'raw')
}
