% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_data.R
\name{site_data}
\alias{site_data}
\title{Simple Site Data}
\usage{
site_data(site_metadata, monitors, site_list, sitecode_col = "aqs_sitecode")
}
\arguments{
\item{site_metadata}{data.frame Site metadata}

\item{monitors}{data.frame Monitors metadata}

\item{site_list}{character vecotr List of program sites to be included in
data frame}

\item{sitecode_col}{string (optional) Quoted name of column with aqs site
codes in all 3 of the above files. Default = 'aqs_sitecode'.}
}
\value{
data.frame Site data ready for plotting
}
\description{
Generates site data for plotting on map.
}
\details{
Requires site metadata and monitors metadata, as read-in by the
\code{import_metadata()} function.
}
