% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_method_summary_data.R
\name{site_method_summary_data}
\alias{site_method_summary_data}
\title{Site-method summary data}
\usage{
site_method_summary_data(sample_data, methods_list)
}
\arguments{
\item{sample_data}{data.frame Sample data data.frame}

\item{methods_list}{data.frame Methods metadata}
}
\value{
data.frame Summary statistics per site-year-parameter-method-poc
}
\description{
Returns summary statistics for hourly observations for each site/method
including, min, max, mean, count valid, etc.
}
\examples{
site_method_summary_data(sample_data, methods_list)
}
