% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_annual_data.R
\name{tidy_annual_data}
\alias{tidy_annual_data}
\title{Tidy annual data}
\usage{
tidy_annual_data(annual_data_adj_poc)
}
\arguments{
\item{annual_data_adj_poc}{data.frame Annual data passed through \code{adjust_annual_poc()}}
}
\value{
data.frame
}
\description{
Filters data for validity_indicator == 'Y' and event_type == 'No Events'.
}
\details{
Filters annual data per aqs_sitecode, parameter, method, poc,
sample_duration and pollutant_standard to only groups with data for
more than 75\% of total years available.
}
\examples{
annual_data_adj_poc <- adjust_annual_poc(annual_data)
tidy_annual_data <- tidy_annual_data(annual_data_adj_poc)
}
