% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_data}
\alias{ts_data}
\title{Generate time series dataframe}
\usage{
ts_data(sample_data, method_mdl_summary, duration)
}
\arguments{
\item{sample_data}{data.frame sampleData data.frame}

\item{method_mdl_summary}{Detection limit summary data as generated by the
\code{method_mdl_summary()} function.}

\item{duration}{string The sample duration to include in the dataframe as it
appears in the sample_duration column (e.g.
'1 HOUR', '24-HOUR').}
}
\value{
data.frame Tidied dataframe with measurement data for the specified
duration per date-time value across parameters.
}
\description{
Requires raw data (data) and detection limits summary data, generated
with the \code{detection_limit_data()} function.
}
\details{
Returns a tidied dataframe with time series data for all non-wind parameters.
}
\examples{
method_mdl_summary <- method_mdl_summary(sample_data)
ts_1hr_data <- ts_data(sample_data, method_mdl_summary, duration = '1 HOUR')
}
