% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_parameters.R
\name{ts_parameters}
\alias{ts_parameters}
\title{TS parameters}
\usage{
ts_parameters(ts_data)
}
\arguments{
\item{ts_data}{data.frame Time series data generated by \code{ts_data()}}
}
\value{
data.frame List of sites and parameters in a time series dataframe
}
\description{
Create data frame listing unique parameters per site from a
time series dataframe (generated by \code{ts_data()}).
}
\examples{
method_mdl_summary <- method_mdl_summary(sample_data)
ts_1hr_data <- ts_data(sample_data, method_mdl_summary, duration = '1 HOUR')
ts_parameters(ts_1hr_data)
}
