% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_data.R
\name{wind_data}
\alias{wind_data}
\title{Tidy & extract wind data}
\usage{
wind_data(sample_data)
}
\arguments{
\item{sample_data}{data.frame Hourly sample data. Should include
met parameters for wind direction and wind
speed.}
}
\value{
data.frame Tidied wind data with columns for wind direction,
wind speed, sitecode, and datetime (posixct).
}
\description{
Extract wind data from sample data and returns data frame with columns for
date-time, formatted as posixct, aqs sitecode, wind direction, & wind speed.
}
